Add user endpoint

+++ db/tables.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS public.user;
 
 CREATE TABLE public.user 
 (
-    id SERIAL PRIMARY KEY,
+    id NUMERIC PRIMARY KEY,
     name VARCHAR(50) NOT NULL,
     email VARCHAR(50) NOT NULL,
     password VARCHAR(50) NOT NULL
